Write failed operations to an actions_file
See https://github.com/bulletmark/edir/issues/13

If the requested operations cannot be executed for some reason (for
example missing permissions, insufficient disk space, etc.) the changes
are lost. This commit writes them to an "actions file" instead to
reapply them later.

A new option "-i" is introduced to read the actions from an actions file
and reapply them instead of opening an editor for interactive use.

The actions file is human readable and may be edited by the user.
However, care must be taken to not change the order of the actions as
they may depend on another (for example a file should be copied before
being renamed, otherwise the copy operation will fail).