Provide a vim ftplugin for the edir buffer
The filetype is detected via file name (must start with 'edir' or 'dir')
and the content (the first 3 non-blank lines must be formatted like an
edir buffer).

A syntax file is provided to visually differentiate between the counter
and the actual file name.

The original filename and additional info (from 'ls') is displayed as
virtual text for the line the cursor is on (currently only for neovim).

The 'tabstop' setting is adjusted to the optimal value to separate the
counter and the file names.